feat: Switched the array to a List
Used Java streams to pipeline the given numbers from the user, using Java.util.function IntFunction() which accepts an int as an argument and returns an int as a result. also have accounted for the values that are negative and null values. Still need to figure out how I will keep the user on the same value so they do not have to restart the whole question over and over again.