feat: fixed a bunch of the code
Stil need to add the remaining comments but for the most part this is done, and it is working really well - Still don't take my word for it 100% because it is 2 am right now and I need to go to bed. Just know that this is working almost 100% and it's very nice.